+++ tests/SaveDeletedByUserTest.php
@@ -7,7 +7,6 @@ use Illuminate\Database\Eloquent\SoftDeletes;
 use TestMonitor\Accountable\Test\Models\User;
 use TestMonitor\Accountable\Test\Models\Record;
 use TestMonitor\Accountable\Traits\Accountable;
-use TestMonitor\Accountable\AccountableSettings;
 use TestMonitor\Accountable\Test\Models\SoftDeletableUser;
 
 class SaveDeletedByUserTest extends TestCase
@@ -16,11 +15,6 @@ class SaveDeletedByUserTest extends TestCase
      * @var \TestMonitor\Accountable\Test\Models\Record
      */
     protected $record;
-
-    /**
-     * @var AccountableSettings
-     */
-    protected $config;
 
     public function setUp(): void
     {
@@ -31,8 +25,6 @@ class SaveDeletedByUserTest extends TestCase
         $this->record = new class() extends Record {
             use Accountable, SoftDeletes;
         };
-
-        $this->config = app()->make(AccountableSettings::class);
     }
 
     #[Test]
@@ -47,7 +39,6 @@ class SaveDeletedByUserTest extends TestCase
 
         $this->assertEquals($record->deleted_by_user_id, User::first()->id);
         $this->assertEquals($record->deleter->name, User::first()->name);
-        $this->assertEquals($record->deletedBy->name, User::first()->name);
         $this->assertInstanceOf(get_class(User::first()), $record->deleter);
     }
 
@@ -105,7 +96,7 @@ class SaveDeletedByUserTest extends TestCase
 
         $anonymous = ['name' => 'Neville the Fat Hamster'];
 
-        $this->config->setAnonymousUser($anonymous);
+        accountable()->setAnonymousUser($anonymous);
 
         $this->assertNull($record->deleted_by_user_id);
         $this->assertInstanceOf(User::class, $record->deleter);
@@ -133,7 +124,7 @@ class SaveDeletedByUserTest extends TestCase
     #[Test]
     public function it_will_save_a_specified_user_as_deleter_when_disabling_accountable()
     {
-        $this->config->disable();
+        accountable()->disable();
 
         $this->actingAs(User::all()->first());